+++ src/Atoms/Table/Thead/Thead.types.ts
@@ -1,5 +1,6 @@
 export type Props = {
   stickyHeader?: boolean;
+  children?: React.ReactNode;
 };
 
 export type TheadComponent = React.FunctionComponent<Props>;