+++ src/Molecules/FlexTable/Cell/Cell.types.ts
@@ -43,6 +43,7 @@ type TextWrapperProps = {
   truncate?: boolean;
   className?: string;
   weight?: string;
+  children?: React.ReactNode;
 };
 
 export type TextWrapperComponent = React.FC<TextWrapperProps>;