+++ src/Organisms/ShowMore/ShowMore.types.ts
@@ -8,4 +8,5 @@ export type Props = {
   showMoreText: string;
   onShowMore?: React.MouseEventHandler;
   linesToClamp?: number;
+  children?: React.ReactNode;
 };