+++ src/Organisms/StatusModal/StatusModal.types.ts
@@ -1,4 +1,9 @@
+import { Theme } from '../../theme/theme.types';
+
 export type Status = 'SUCCESS' | 'ERROR' | 'WARNING';
+
+type Colors = Theme['color'];
+type ColorFn = (t: Theme) => Colors['cta'] | Colors['negative'] | Colors['functionRed'];
 
 type Option = {
   status?: Status;
@@ -14,4 +19,6 @@ export type Props = {
   options: Option;
   id: string;
   showClose?: boolean;
+  fixedBottomMobile?: boolean;
+  confirmColor?: ColorFn;
 };